third_party/WebKit/LayoutTests/virtual/threaded/animations/composited-filter-webkit-filter.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
@keyframes test {
  from {
    -webkit-filter: saturate(0);
    filter: none;
  }
  to {
    -webkit-filter: saturate(1);
    filter: none;
  }
}
#target {
  animation: test 1e10s;
}
</style>
<div id="target"></div>
<script>
var asyncHandle = async_test('-webkit-filter + filter animations get composited.');
requestAnimationFrame(_ => {
  requestAnimationFrame(_ => {
    asyncHandle.step(_ => {
      assert_true(internals.isCompositedAnimation(target.getAnimations()[0]));
    });
    asyncHandle.done();
  });
});
</script>
